+++ app/Http/Controllers/Admin/StudentLifeController.php
@@ -18,52 +18,79 @@ class StudentLifeController extends Controller
         return view('admin.studentlife.edit', compact('studentLife'));
     }
 
-    public function update(UpdateStudentLifeRequest $request, StudentLife $studentLife)
+    public function update(UpdateStudentLifeRequest $request, $studentLife)
     {
+
+        $studentLife = StudentLife::find($studentLife);
 
         abort_if(Gate::denies('content_management_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $file1 = $request->file('pic1');
         $file2 = $request->file('pic2');
         $file3 = $request->file('pic3');
+        $file4 = $request->file('main_img');
 
 
         if ($file1 != ''){
-            $path = public_path() . '/images';
+            $path = public_path() . '/images/';
 
-            if ($studentLife->pic1 != '' && $studentLife->pic1 != null){
-                $file_old1 = $path.$studentLife->pic1;
-                unlink($file_old1);
+            if ($path.$studentLife->pic1){
+                if ($studentLife->pic1 != '' && $studentLife->pic1 != null){
+                    $file_old1 = $path.$studentLife->pic1;
+                    unlink($file_old1);
+                }
             }
+            
             $pic1 = time().'.'. $file1->getClientOriginalName();
             $path = $file1->storeAs('public/images', $pic1);
             $request->pic1->move(public_path('images'), $pic1);
-            $request->pic1 = $path;
+            $request->pic1 = $pic1;
         }
         if ($file2 != ''){
-            $path = public_path() . '/images';
+            $path = public_path() . '/images/';
 
-            if ($studentLife->pic2 != '' && $studentLife->pic2 != null){
-                $file_old2 = $path.$studentLife->pic2;
-                unlink($file_old2);
+            if ($path.$studentLife->pic2){
+                if ($studentLife->pic2 != '' && $studentLife->pic2 != null){
+                    $file_old2 = $path.$studentLife->pic2;
+                    unlink($file_old2);
+                }
             }
+            
             $pic2 = time().'.'. $file2->getClientOriginalName();
             $path = $file2->storeAs('public/images', $pic2);
             $request->pic2->move(public_path('images'), $pic2);
-            $request->pic2 = $path;
+            $request->pic2 = $pic2;
         }
 
         if ($file3 != ''){
-            $path = public_path() . '/images';
+            $path = public_path() . '/images/';
 
-            if ($studentLife->pic3 != '' && $studentLife->pic3 != null){
-                $file_old3 = $path.$studentLife->pic3;
-                unlink($file_old3);
+            if ($path.$studentLife->pic3){
+                if ($studentLife->pic3 != '' && $studentLife->pic3 != null){
+                    $file_old3 = $path.$studentLife->pic3;
+                    unlink($file_old3);
+                }
             }
+            
             $pic3 = time().'.'. $file3->getClientOriginalName();
             $path = $file3->storeAs('public/images', $pic3);
             $request->pic3->move(public_path('images'), $pic3);
-            $request->pic3 = $path;
+            $request->pic3 = $pic3;
+        }
+        if ($file4 != ''){
+            $path = public_path() . '/images/';
+
+            if ($path.$studentLife->main_img){
+                if ($studentLife->main_img != '' && $studentLife->main_img != null){
+                    $file_old4 = $path.$studentLife->main_img;
+                    unlink($file_old4);
+                }
+            }
+
+            $main_img = time().'.'. $file4->getClientOriginalName();
+            $path = $file4->storeAs('public/images', $main_img);
+            $request->main_img->move(public_path('images'), $main_img);
+            $request->main_img = $main_img;
         }
 
 
@@ -82,7 +109,11 @@ class StudentLifeController extends Controller
             $index->pic3 = $pic3;
             $index->save();
         }
+        if ($file4 != '') {
+            $index->main_img = $main_img;
+            $index->save();
+        }
 
-        return redirect()->route('admin.preschool.index');
+        return redirect()->route('admin.studentlife.index');
     }
 }
